refactor: split services into subpackages
src/main/java/com/safetynet/safetynetalertsapi/services/finders/MedicalRecordFinder.java
src/main/java/com/safetynet/safetynetalertsapi/services/finders/PersonFinder.java
src/main/java/com/safetynet/safetynetalertsapi/services/finders/FireStationFinder.java

Move all Finder services to the new package services.finder

src/main/java/com/safetynet/safetynetalertsapi/services/mappers/PersonMapper.java

Move PersonMapper services to the new package services.mapper

src/main/java/com/safetynet/safetynetalertsapi/services/collectionutils/PersonFilterService.java

Create package services.collectionutils for collection-related utilities
(e.g. filtering, sorting). Move PersonFilterService to package
services.collectionutils.
